services/vector-db/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
faiss-cpu==1.7.4
numpy==1.26.2
